backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
PyPDF2
passlib[bcrypt]
python-jose[cryptography]
pymongo 